requirements.txt
google-generativeai
flask
pandas
scikit-learn
pytesseract
pdfplumber